% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_assign_obj.R
\name{clip_assign_obj}
\alias{clip_assign_obj}
\title{Assign objects from clipboard written with clip_write_obj}
\usage{
clip_assign_obj(envir = .GlobalEnv)
}
\arguments{
\item{envir}{environment in which to assign the objects, defaults to global}
}
\value{
NULL, invisibly
}
\description{
Assign objects from clipboard written with clip_write_obj
}
