% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_read1_obj.R
\name{clip_read1_obj}
\alias{clip_read1_obj}
\title{Read first object from clipboard written with clip_write_obj}
\usage{
clip_read1_obj()
}
\value{
object from clipboard (could be of any type)
}
\description{
Read first object from clipboard written with clip_write_obj
}
