% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_read_obj.R
\name{clip_read_obj}
\alias{clip_read_obj}
\title{Read objects from clipboard written with clip_write_obj}
\usage{
clip_read_obj()
}
\value{
list of objects from clipboard
}
\description{
Read objects from clipboard written with clip_write_obj
}
