% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_write_obj.R
\name{clip_write_obj}
\alias{clip_write_obj}
\title{Write object(s) to clipboard}
\usage{
clip_write_obj(...)
}
\arguments{
\item{...}{Objects to write, optionally with names}
}
\value{
objects passed in as a list, invisibly
}
\description{
Write object(s) to clipboard
}
